Allow empty lines between let bindings.
This no longer forces the lexer to see every 'let' after a newline as a
top-level let definition, which was previously necessary to allow
indenting top-level definitions in the first place.

A 'let' is now considered top-level if it has the indentation of a
top-level declaration.  This means that the lexer partly determines a
token based on the number of spaces in front of it!

One flip side is that the lexing gets a bit slower, since for every 'let' token
it needs to traverse the program backwards -- possibly to the very beginning of
the file -- to see if it is enclosed in a module/open block.  This is noticable
when indenting a region of ~100 lines and up (on my high-end laptop!).

We solve this by keeping a little cache state around *only* when indenting
regions.  This seems to work, and is pretty simple, but we'll see if the added
complexity outweighs the benefits (it's easy to disable again).

Also add test case.